requirements.txt
aiogram==3.4.1
aiohttp==3.9.3
beautifulsoup4==4.12.3
httpx==0.27.0
instaloader==4.10.3
numpy==1.26.4
Pillow==10.2.0
python-dotenv==1.0.1
Requests==2.31.0
selenium==4.18.1
selenium_stealth==1.0.6
webdriver_manager==4.0.1
